requirements.txt
matplotlib==3.8.4
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
scikit-learn==1.4.2
torch==2.2.2
tqdm==4.66.4
wandb==0.17.2